Organize files, start using renderer.
