--- chords_renderer.js
@@ -1,10 +1,0 @@
-'use strict';
-
-function renderChords(str, opts) {
-  return `<b>${str}</b>`;
-}
-
-module.exports = {
-  'lang': 'chords',
-  'callback': renderChords
-};
