+++ index.js
@@ -1,6 +1,6 @@
 'use strict';
 const abc = require('./abc_renderer.js');
-const chords = require('./chords_renderer.js');
+const chords = require('./renderers/chords_renderer.js');
 
 function MarkdownMusic(md, musicOpts) {
   md.highlightRegistry = {};
