+++ instruments/chord.js
